refactor: 클래스명 변경으로 인한 코드 수정
