+++ src/main/java/mju/chatuniv/member/service/service/MemberService.java
@@ -2,9 +2,9 @@ package mju.chatuniv.member.service.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import mju.chatuniv.board.controller.dto.BoardResponse;
 import mju.chatuniv.board.domain.Board;
 import mju.chatuniv.board.domain.BoardRepository;
+import mju.chatuniv.board.domain.dto.BoardResponse;
 import mju.chatuniv.chat.domain.chat.Chat;
 import mju.chatuniv.chat.domain.chat.ChatRepository;
 import mju.chatuniv.comment.domain.CommentRepository;
